+++ bin/c-mongo-download.js
@@ -9,13 +9,15 @@ const { loadConfig, reportError } = require('./lib/c');
 // The basic program, which uses sub-commands.
 program
     .arguments('[env]')
+    .arguments('<collection>')
+    .description('Download all or a specific collection from a Mongo database. If you don\'t provide a collection, all will be downloaded.')
     .parse(process.argv);
 
 if (!program.args.length) {
     return reportError(new Error('Please provide an environment'), program);
 }
 
-const [env] = program.args;
+const [env, collection] = program.args;
 
 if (env.toLowerCase() === 'local') {
     return reportError(new Error('You cannot download the local database'), program);
@@ -25,12 +27,17 @@ loadConfig(`mongo.${env}`)
     .then((db) => {
 
         let ssl = '';
+        let collectionArg = '';
 
         if ((typeof db.ssl === 'undefined') ? true : db.ssl) {
             ssl = '--ssl --sslAllowInvalidCertificates';
         }
 
-        return spawn(`docker run -it -v ${process.cwd()}/data:/data --rm mongo:latest mongodump ${ssl} -h ${db.host}:${db.port} -u ${db.user} -p ${db.password} -d ${db.name} -o data`, {
+        if (typeof collection !== 'undefined') {
+            collectionArg = `-c ${collection}`;
+        }
+
+        return spawn(`docker run -it -v ${process.cwd()}/data:/data --rm mongo:latest mongodump ${ssl} -h ${db.host}:${db.port} -u ${db.user} -p ${db.password} -d ${db.name} ${collectionArg} -o data`, {
             shell: true,
             stdio: 'inherit',
         });