+++ bin/c-mongo-import.js
@@ -9,13 +9,15 @@ const { loadConfig, reportError } = require('./lib/c');
 // The basic program, which uses sub-commands.
 program
     .arguments('[env]')
+    .arguments('<collection>')
+    .description('Import all or a specific collection from a Mongo database. If you don\'t provide a collection, all will be imported.')
     .parse(process.argv);
 
 if (!program.args.length) {
     return reportError(new Error('Please provide an environment'), program);
 }
 
-const [env] = program.args;
+const [env, collection] = program.args;
 
 if (env.toLowerCase() === 'local') {
     return reportError(new Error('You cannot download the local database'), program);
@@ -27,6 +29,14 @@ loadConfig('mongo')
         const db = mongo[`${env}`];
         const localDb = mongo.local;
 
+        // Default to importing all collections.
+        let collectionArg = `--nsInclude '${db.name}.*'`;
+
+        // Otherwise, import a specific collection only.
+        if (typeof collection !== 'undefined') {
+            collectionArg = `--nsInclude '${db.name}.${collection}'`;
+        }
+
         if (!localDb) {
             return reportError(new Error('Could not find the local db'), program);
         }
@@ -35,7 +45,7 @@ loadConfig('mongo')
             return reportError(new Error(`Could not find the ${env} db`), program);
         }
 
-        return spawn(`docker run -it -v ${process.cwd()}/data/${db.name}:/data/${db.name} --add-host ${localDb.host}:$(c hosts get -n ${localDb.host}) --rm mongo:latest mongorestore --noIndexRestore --drop -h ${localDb.host}:${localDb.port} -d ${localDb.name} data/${db.name}`, {
+        return spawn(`docker run -it -v ${process.cwd()}/data/${db.name}:/data/${db.name} --add-host ${localDb.host}:$(c hosts get -n ${localDb.host}) --rm mongo:latest mongorestore --noIndexRestore --drop -h ${localDb.host}:${localDb.port} ${collectionArg} data/`, {
             shell: true,
             stdio: 'inherit',
         });