+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 This file is a history of the changes made to @idearium/cli.
 
+## Unreleased (30 April 2019)
+
+- Improved commands.
+
+### Improvements
+
+- `c mongo download`, `c mongo import` and `c mongo sync` now support a `collection` argument to act only on a specific collection.
+
 ## v2.1.0 (2 April 2019)
 
 - New workflow.